backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.7
Django==5.2.13
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
exceptiongroup==1.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
kombu==5.6.2
numpy==2.2.6
packaging==26.1
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
pyparsing==3.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.4.0
requests==2.33.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
